refactor: split sources panel module
